requirements.txt
numpy==1.20.1
streamlit
seaborn==0.11.1
matplotlib==3.4.1
plotly==4.14.3
pandas==1.2.3
joblib==0.16.0
scikit-learn
Pillow
